tests/wpt/web-platform-tests/app-history/per-entry-events/dispose-same-document-replace-with-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async (t) => {
  // Wait for after the load event so that we're definitely testing intentional,
  // navigate()-caused replacement and not the replacement that happens
  // automatically before the load event completes.
  await new Promise(r => window.onload = () => t.step_timeout(r, 0));

  const entriesBefore = appHistory.entries();
  const currentBefore = appHistory.current;

  let disposeCalled = false;
  appHistory.current.ondispose = t.step_func(e => {
    disposeCalled = true;

    assert_equals(e.constructor, Event);
    assert_equals(e.bubbles, false);
    assert_equals(e.cancelable, false);
    assert_equals(e.composed, false);

    assert_not_equals(appHistory.current, currentBefore);
    assert_array_equals(appHistory.entries(), [appHistory.current]);
    assert_equals((new URL(appHistory.current.url)).search, "?replacement");
    assert_equals(appHistory.transition.navigationType, "replace");
    assert_equals(appHistory.transition.from, entriesBefore[0]);
    assert_equals(location.search, "?replacement");
  });

  appHistory.addEventListener("navigate", e => e.transitionWhile(Promise.resolve()));

  appHistory.navigate("?replacement", { replace: true });
  assert_true(disposeCalled);
}, "dispose events when doing a same-document replace using appHistory.navigate() and transitionWhile()");
</script>
